+++ src/ONTraC/analysis/spatial.py
@@ -53,6 +53,7 @@ def plot_cell_type_composition_dataset(ana_data: AnaData) -> Optional[Tuple[plt.
     fig.tight_layout()
     if ana_data.options.output is not None:
         fig.savefig(f'{ana_data.options.output}/cell_type_composition.pdf', transparent=True)
+        plt.close(fig)
         return None
     else:
         return fig, axes
@@ -257,6 +258,7 @@ def plot_niche_NT_score_dataset(ana_data: AnaData) -> Optional[Tuple[plt.Figure,
     fig.tight_layout()
     if ana_data.options.output is not None:
         fig.savefig(f'{ana_data.options.output}/niche_NT_score.pdf', transparent=True)
+        plt.close(fig)
         return None
     else:
         return fig, axes
@@ -346,6 +348,7 @@ def plot_cell_NT_score_dataset(ana_data: AnaData) -> Optional[Tuple[plt.Figure, 
     fig.tight_layout()
     if ana_data.options.output is not None:
         fig.savefig(f'{ana_data.options.output}/cell_NT_score.pdf', transparent=True)
+        plt.close(fig)
         return None
     else:
         return fig, axes